+++ packages/react-html/src/components/tests/AppleHomeScreen.test.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
-import {mount, Root} from '@shopify/react-testing';
-import {Props} from '@shopify/useful-types';
+import {mount} from '@shopify/react-testing';
 
 import Link from '../Link';
 import Meta from '../Meta';
@@ -10,24 +9,18 @@ import AppleHomeScreen, {IconSize} from '../AppleHomeScreen';
 describe('<AppleHomeScreen />', () => {
   it('renders a <Meta /> with `name="apple-mobile-web-app-capable" content="yes"` by default', () => {
     const appleHomeScreen = mount(<AppleHomeScreen />);
-
-    expect(
-      findMeta(appleHomeScreen, {
-        name: 'apple-mobile-web-app-capable',
-        content: 'yes',
-      }),
-    ).not.toBeNull();
+    expect(appleHomeScreen).toContainReactComponent(Meta, {
+      name: 'apple-mobile-web-app-capable',
+      content: 'yes',
+    });
   });
 
   it('renders a <Meta /> with `name="apple-mobile-web-app-status-bar-style" content="black"` by default', () => {
     const appleHomeScreen = mount(<AppleHomeScreen />);
-
-    expect(
-      findMeta(appleHomeScreen, {
-        name: 'apple-mobile-web-app-status-bar-style',
-        content: 'black',
-      }),
-    ).not.toBeNull();
+    expect(appleHomeScreen).toContainReactComponent(Meta, {
+      name: 'apple-mobile-web-app-status-bar-style',
+      content: 'black',
+    });
   });
 
   it('renders a <Link /> for each item in the icons prop', () => {
@@ -49,34 +42,21 @@ describe('<AppleHomeScreen />', () => {
     const appleHomeScreen = mount(<AppleHomeScreen icons={icons} />);
 
     icons.forEach(({size, url}, index) => {
-      expect(appleHomeScreen.findAll(Link)[index].props).toMatchObject({
+      expect(appleHomeScreen.findAll(Link)[index]).toHaveReactProps({
         sizes: `${size}x${size}`,
         href: url,
       });
     });
   });
+
   it('renders a <Link /> for the startUpImage prop', () => {
     const startUpImage = 'path/to/startup.png';
     const appleHomeScreen = mount(
       <AppleHomeScreen startUpImage={startUpImage} />,
     );
 
-    expect(appleHomeScreen.find(Link)!.props).toMatchObject({
+    expect(appleHomeScreen).toContainReactComponent(Link, {
       href: startUpImage,
     });
   });
 });
-
-export function findMeta(wrapper: Root<any>, props: Props<typeof Meta>) {
-  return wrapper.findWhere(element => {
-    if (!element.is(Meta)) {
-      return false;
-    }
-
-    const propsArray: (keyof typeof props)[] = Object.keys(props) as any;
-    return (
-      propsArray.filter(prop => element.prop(prop) === props[prop]).length ===
-      propsArray.length
-    );
-  });
-}